Consolidate SwiftUI preview to a single renderer
`AccessibilitySnapshotView` (formerly `PreParsedAccessibilitySnapshotView`)
is now the only public SwiftUI renderer. Live parsing for
`View.accessibilityPreview()` lives in a private `LiveAccessibilityPreview`
wrapper that captures the parsed hierarchy and passes it through, so
`showContainers: true` works in #Preview canvases (not just in tests).

Removes the orphan `AccessibilityPreviews/SwiftUIAccessibilitySnapshotView.swift`
(pre-#320 dead code), the old generic `AccessibilitySnapshotView<Content>`
struct, the `SwiftUIAccessibilitySnapshotView` typealias, and the
unreferenced `UIViewWrapper` helpers.
