+++ Sources/AccessibilitySnapshot/AccessibilitySnapshotPreviews/SwiftUIAccessibilitySnapshotView.swift
@@ -2,150 +2,14 @@ import AccessibilitySnapshotCore
 import AccessibilitySnapshotParser
 import SwiftUI
 
-/// A SwiftUI container view that displays a snapshot with accessibility overlays and legend.
+/// A SwiftUI view that displays a rendered snapshot with accessibility overlays and legend.
+///
+/// Callers supply the already-parsed snapshot data (image, markers, optional hierarchy). The
+/// live-parsing entry point used by `View.accessibilityPreview()` is a private wrapper that
+/// runs the parser and then constructs this view; tests construct it directly from
+/// `SwiftUIAccessibilitySnapshotContainerView`.
 @available(iOS 16.0, *)
-public struct AccessibilitySnapshotView<Content: View>: View {
-    private let content: Content
-    private let configuration: AccessibilitySnapshotConfiguration
-    private let palette: ColorPalette
-    private let renderSize: CGSize
-
-    @State private var markers: [AccessibilityMarker] = []
-    @State private var snapshotImage: UIImage?
-    @State private var parseError: Error?
-
-    public init(
-        @ViewBuilder content: () -> Content,
-        configuration: AccessibilitySnapshotConfiguration = .init(viewRenderingMode: .drawHierarchyInRect),
-        palette: ColorPalette = .default,
-        renderSize: CGSize? = nil
-    ) {
-        self.content = content()
-        self.configuration = configuration
-        self.palette = palette
-        self.renderSize = renderSize ?? UIScreen.main.bounds.size
-    }
-
-    private var showUserInputLabels: Bool {
-        configuration.inputLabelDisplayMode != .never
-    }
-
-    private var showUnspokenTraits: Bool {
-        configuration.showsUnspokenTraits
-    }
-
-    public var body: some View {
-        VStack(spacing: 0) {
-            if let snapshotImage = snapshotImage {
-                snapshotWithOverlays(image: snapshotImage)
-            } else if let parseError = parseError {
-                errorView(error: parseError)
-            }
-
-            LegendView(
-                markers: markers,
-                palette: palette,
-                showUserInputLabels: showUserInputLabels,
-                showUnspokenTraits: showUnspokenTraits
-            )
-            .frame(width: renderSize.width)
-        }
-        .onAppear {
-            if snapshotImage == nil {
-                parseAccessibility()
-            }
-        }
-        .background(Color(white: 0.9))
-    }
-
-    // MARK: - Private Views
-
-    private func snapshotWithOverlays(image: UIImage) -> some View {
-        SnapshotOverlayView(
-            snapshotImage: image,
-            markers: markers,
-            palette: palette,
-            renderSize: renderSize,
-            activationPointDisplayMode: configuration.activationPointDisplayMode
-        )
-    }
-
-    @ViewBuilder
-    private func errorView(error: Error) -> some View {
-        VStack {
-            Image(systemName: "exclamationmark.triangle.fill")
-            Text(error.localizedDescription)
-        }
-        content
-    }
-
-    // MARK: - Private Methods
-
-    private func parseAccessibility() {
-        let adjustedContent = content
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-            .edgesIgnoringSafeArea(.all)
-
-        let hostingController = UIHostingController(rootView: adjustedContent)
-        hostingController.view.frame = CGRect(origin: .zero, size: renderSize)
-
-        // UIViewRepresentable views (e.g. PathShapeUIView) report accessibilityPath and
-        // accessibilityFrame in screen coordinates via `convert(_:to: nil)`, which requires
-        // the view to be installed in a UIWindow. Without one, every view reports (0,0) and
-        // the paths collapse to the origin.
-        let window = UIWindow(frame: CGRect(origin: .zero, size: renderSize))
-        window.rootViewController = hostingController
-        window.makeKeyAndVisible()
-        hostingController.view.layoutIfNeeded()
-
-        defer {
-            window.isHidden = true
-            window.rootViewController = nil
-        }
-
-        do {
-            snapshotImage = try hostingController.view.renderToImage(
-                configuration: configuration.rendering
-            )
-
-            let parser = AccessibilityHierarchyParser()
-            markers = parser.parseAccessibilityHierarchy(
-                in: hostingController.view,
-                rotorResultLimit: configuration.rotors.resultLimit
-            ).flattenToElements()
-        } catch {
-            parseError = error
-        }
-    }
-}
-
-/// Backwards compatibility alias.
-@available(iOS 16.0, *)
-public typealias SwiftUIAccessibilitySnapshotView<Content: View> = AccessibilitySnapshotView<Content>
-
-// MARK: - UIView Wrapper
-
-@available(iOS 16.0, *)
-public extension AccessibilitySnapshotView where Content == UIViewWrapper {
-    /// Creates a snapshot view wrapping a UIView.
-    init(
-        containedView: UIView,
-        configuration: AccessibilitySnapshotConfiguration = .init(viewRenderingMode: .drawHierarchyInRect),
-        palette: ColorPalette = .default,
-        renderSize: CGSize? = nil
-    ) {
-        content = UIViewWrapper(view: containedView)
-        self.configuration = configuration
-        self.palette = palette
-        self.renderSize = renderSize ?? containedView.bounds.size
-    }
-}
-
-// MARK: - Pre-parsed Snapshot View
-
-/// A SwiftUI view that displays a pre-rendered snapshot with accessibility overlays and legend.
-@available(iOS 16.0, *)
-public struct PreParsedAccessibilitySnapshotView: View {
+public struct AccessibilitySnapshotView: View {
     private let snapshotImage: UIImage
     private let markers: [AccessibilityMarker]
     private let hierarchy: [AccessibilityHierarchy]
@@ -332,24 +196,4 @@ private struct SnapshotOverlayView: View {
             return false
         }
     }
-}
-
-/// A SwiftUI wrapper for UIView.
-@available(iOS 16.0, *)
-public struct UIViewWrapper: View {
-    let view: UIView
-
-    public var body: some View {
-        UIViewWrapperRepresentable(view: view)
-    }
-}
-
-private struct UIViewWrapperRepresentable: UIViewRepresentable {
-    let view: UIView
-
-    func makeUIView(context: Context) -> UIView {
-        view
-    }
-
-    func updateUIView(_ uiView: UIView, context: Context) {}
 }
